Extract generateMultipleFunctionParams to own file in src/aws
