+++ bin/deploy.js
@@ -17,28 +17,6 @@ const lambdaRoleName = "lambda_basic_execution";
 const statesRoleName = "stepFunctions_basic_execution";
 const basenamesAndZipBuffers = getBasenamesAndZipBuffers();
 
-const generateFunctionParams = (basename, zipBuffer, role) => {
-  return {
-    Code: {
-      ZipFile: zipBuffer,
-    },
-    FunctionName: basename,
-    Handler: `${basename}.handler`,
-    Role: role.Role.Arn,
-    Runtime: "nodejs12.x",
-  };
-};
-
-const generateMultipleFunctionParams = async (
-  basenamesAndZipBuffers,
-  roleName
-) => {
-  const role = await iam.getRole({ RoleName: roleName }).promise();
-
-  return basenamesAndZipBuffers.map(({ basename, zipBuffer }) => {
-    return generateFunctionParams(basename, zipBuffer, role);
-  });
-};
 const createLambdaFunctions = (allParams) => {
   const createFunctionPromises = allParams.map((params) =>
     retryAsync(() => lambda.createFunction(params).promise(), 5, 7000, 0.6)
